+++ package.json
@@ -18,10 +18,10 @@
   },
   "dependencies": {
     "@automerge/automerge": "3.1.2",
-    "@automerge/automerge-repo": "2.3.0",
-    "@automerge/automerge-repo-network-websocket": "2.3.0",
-    "@automerge/automerge-repo-react-hooks": "2.3.0",
-    "@automerge/automerge-repo-storage-indexeddb": "2.3.0",
+    "@automerge/automerge-repo": "2.2.0",
+    "@automerge/automerge-repo-network-websocket": "2.2.0",
+    "@automerge/automerge-repo-react-hooks": "2.2.0",
+    "@automerge/automerge-repo-storage-indexeddb": "2.2.0",
     "@mantine/core": "8.3.1",
     "@mantine/dates": "8.3.1",
     "@mantine/dropzone": "8.3.1",
