+++ src/components/docContext/hooks/useDocContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { DocContext } from "@/components/docContext/DocContext.tsx";
+
+export function useDocContext() {
+  return useContext(DocContext);
+}
