+++ src/components/serversContext/ServersContextProvider.tsx
@@ -1,0 +1,174 @@
+import { useCallback, useEffect, useState } from "react";
+import { Repo } from "@automerge/automerge-repo";
+import { IndexedDBStorageAdapter } from "@automerge/automerge-repo-storage-indexeddb";
+import { openDeleteConfirmModal } from "@/components/modals/modals";
+import { STORAGE_KEYS } from "@/lib/constants/localStorageKeys";
+import { TDocType } from "@/types/schema";
+import { useInitServerConnections } from "./hooks/useInitServerConnections"; // global handlers need to be a singleton
+import { useSyncServersOnStorage } from "./hooks/useSyncServersOnStorage";
+import { ServersContext, serversHandler } from "./serversContext.tsx";
+import {
+  REPOSITORY_TYPE,
+  SERVER_STATUS,
+  ServersList,
+  StorageServersConfig,
+  TServersProviderProps,
+  YtServer,
+} from "./types";
+
+export const ServersContextProvider = ({ children }: TServersProviderProps) => {
+  const [servers, setServers] = useState<ServersList>({});
+  const [isServerLoadedFromStorage, setIsServerLoadedFromStorage] =
+    useState(false);
+
+  // On server list update save localstroerage
+  useSyncServersOnStorage(servers, !isServerLoadedFromStorage);
+  useInitServerConnections(servers, setServers);
+
+  const addServer = useCallback((id: string, repo: YtServer) => {
+    setServers((currentServers) => {
+      const newServers = { ...currentServers };
+      newServers[id] = repo;
+      return newServers;
+    });
+  }, []);
+
+  // load initial servers from user's local storage
+  useEffect(() => {
+    if (isServerLoadedFromStorage) {
+      return;
+    }
+
+    // Prepend Offline server
+    const offlineRepoId =
+      localStorage.getItem(STORAGE_KEYS.SERVER_OFFLINE_REPOSITORY_ID) ??
+      undefined;
+
+    const offlineServerInitializer: YtServer = {
+      name: "Offline",
+      id: "offline",
+      type: REPOSITORY_TYPE.offline,
+      status: SERVER_STATUS.CONNECTED,
+      url: "offline",
+      repositoryIds: offlineRepoId ? [offlineRepoId] : [],
+      opened: true, // offline server is always opened at initialization
+    };
+
+    serversHandler["offline"] = new Repo({
+      network: [], // no networks for offline server
+      storage: new IndexedDBStorageAdapter(),
+      sharePolicy: async () => true,
+      enableRemoteHeadsGossiping: true,
+    });
+
+    if (!offlineRepoId) {
+      const docHandle = serversHandler["offline"].create<TDocType>({
+        projects: [],
+        description: "",
+        title: "",
+      });
+
+      offlineServerInitializer.repositoryIds = [docHandle.url];
+      localStorage.setItem(
+        STORAGE_KEYS.SERVER_OFFLINE_REPOSITORY_ID,
+        docHandle.url
+      );
+    }
+
+    // Load config from localStorage
+    const serversFromStorageRaw = localStorage.getItem(
+      STORAGE_KEYS.SERVERS_CONF
+    );
+
+    if (serversFromStorageRaw && serversFromStorageRaw !== "") {
+      const serversFromStorage: StorageServersConfig = JSON.parse(
+        serversFromStorageRaw
+      );
+
+      if (serversFromStorage?.servers) {
+        // Initialize server load and init
+        const serversInitialized: ServersList = {
+          offline: offlineServerInitializer,
+        };
+        serversFromStorage.servers.forEach((server) => {
+          // init the server repo with network adapter
+          serversInitialized[server.id] = {
+            id: server.id,
+            name: server.name,
+            type: REPOSITORY_TYPE.remote,
+            status: SERVER_STATUS.IDLE,
+            url: server.url,
+            repositoryIds: server.repositoryIds,
+            opened: server.opened ?? false, // default to false if not provided
+          };
+        });
+
+        setServers(serversInitialized);
+        setIsServerLoadedFromStorage(true);
+        return;
+      }
+    }
+
+    // No config or invalid config found. Set only offline server.
+    const serversInitialized: ServersList = {
+      offline: offlineServerInitializer,
+    };
+
+    setServers(serversInitialized);
+    setIsServerLoadedFromStorage(true);
+  }, [isServerLoadedFromStorage]);
+
+  const disconnectFromServer = useCallback((serverId: string) => {
+    openDeleteConfirmModal("Disconnect from server?", {
+      confirmButtonLabel: "Disconnect",
+      handleConfirm: () =>
+        setServers((prevServers) => {
+          const newServers = { ...prevServers };
+          newServers[serverId].opened = false;
+
+          return newServers;
+        }),
+    });
+  }, []);
+
+  const removeServer = useCallback((serverId: string) => {
+    openDeleteConfirmModal("Disconnect and remove the server from the list?", {
+      confirmButtonLabel: "Remove",
+      handleConfirm: () =>
+        setServers((prevServers) => {
+          const newServers = { ...prevServers };
+          newServers[serverId].opened = false;
+          delete newServers[serverId];
+          try {
+            serversHandler[serverId].networkSubsystem.disconnect();
+            delete serversHandler[serverId];
+          } catch (error) {
+            console.log("Error while disconnecting from remote server", error);
+          }
+          return newServers;
+        }),
+    });
+  }, []);
+
+  const connectToServer = useCallback((serverId: string) => {
+    setServers((prevServers) => {
+      const newServers = { ...prevServers };
+      newServers[serverId].opened = true;
+      return newServers;
+    });
+  }, []);
+
+  return (
+    <ServersContext.Provider
+      value={{
+        servers,
+        addServer,
+        connectToServer,
+        disconnectFromServer,
+        removeServer,
+      }}
+    >
+      {children}
+    </ServersContext.Provider>
+  );
+};
