+++ src/components/serversContext/hooks/useServersContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { ServersContext } from "@/components/serversContext/serversContext.tsx";
+
+export function useServersContext() {
+  return useContext(ServersContext);
+}
