+++ src/components/shared/RichTextarea/RichTextarea.tsx
@@ -6,7 +6,6 @@ import Image from "@tiptap/extension-image";
 import SubScript from "@tiptap/extension-subscript";
 import Superscript from "@tiptap/extension-superscript";
 import TextAlign from "@tiptap/extension-text-align";
-import Underline from "@tiptap/extension-underline";
 import { useEditor } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import MimeMatcher from "mime-matcher";
@@ -33,7 +32,7 @@ export const RichTextarea: FC<{
     shouldRerenderOnTransaction: true,
     extensions: [
       StarterKit.configure({ link: false }),
-      Underline,
+      // Underline,
       Link,
       Superscript,
       SubScript,
