+++ src/lib/operators/useProjects.ts
@@ -1,6 +1,6 @@
 import { useCallback } from "react";
 import { useDocument } from "@automerge/automerge-repo-react-hooks";
-import { useDocContext } from "@/components/docContext/DocContext";
+import { useDocContext } from "@/components/docContext/hooks/useDocContext.ts";
 import { TDocType } from "@/types/schema";
 import { TOperatorLoaderStatus, TUseProjects } from "./types";
